refactor and improve layout
move more common code into EventBundleNodeBase
fix band by type option
